Fix VecScore scan: int32 → float64 in recall row structs
1 - (embedding <=> $n) returns a double precision float from PostgreSQL.
sqlc incorrectly inferred int32; pgx refuses to scan a float into int64/int32
(no-lossless-conversion error). Changed VecScore to float64 in:
  - RecallArtifactsByVectorRow (knowledge.sql.go)
  - RecallMemoriesByVectorRow  (memories.sql.go)
  - RecallMemoriesByCodeVectorRow (memories.sql.go)
Removed now-redundant float64(...) casts in compat.go.

+++ internal/db/compat.go
@@ -579,7 +579,7 @@ func RecallMemoriesByVector(ctx context.Context, pool *pgxpool.Pool, scopeIDs []
 		mem := memoryFromRecallByVectorRow(r)
 		results[i] = MemoryScore{
 			Memory:   mem,
-			VecScore: float64(r.VecScore),
+			VecScore: r.VecScore,
 		}
 	}
 	return results, nil
@@ -601,7 +601,7 @@ func RecallMemoriesByCodeVector(ctx context.Context, pool *pgxpool.Pool, scopeID
 		mem := memoryFromRecallByCodeVectorRow(r)
 		results[i] = MemoryScore{
 			Memory:   mem,
-			VecScore: float64(r.VecScore),
+			VecScore: r.VecScore,
 		}
 	}
 	return results, nil
@@ -960,7 +960,7 @@ func RecallArtifactsByVector(ctx context.Context, pool *pgxpool.Pool, scopeIDs [
 		art := artifactFromRecallByVectorRow(r)
 		results[i] = ArtifactScore{
 			Artifact: art,
-			VecScore: float64(r.VecScore),
+			VecScore: r.VecScore,
 		}
 	}
 	return results, nil
